requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.26.0
pytest-asyncio==0.23.5 